% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{construct_endpoint}
\alias{construct_endpoint}
\title{Construct the endpoint for the distribution function query}
\usage{
construct_endpoint(path, endpoint_suffix, value)
}
\arguments{
\item{path}{The URL path for an existing distribution.}

\item{endpoint_suffix}{The type of distribution function (pdf, cdf, qf, samples).}

\item{value}{The values to pass in the query (size for samples, x for pdf/cdf, p for qf).}
}
\value{
A string representing the constructed endpoint.
}
\description{
Construct the endpoint for the distribution function query
}
