% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{create_and_query_distribution}
\alias{create_and_query_distribution}
\title{Create a new distribution and query it}
\usage{
create_and_query_distribution(
  api_settings,
  family,
  arguments,
  value,
  endpoint_suffix
)
}
\arguments{
\item{api_settings}{List containing API settings}

\item{family}{The distribution family}

\item{arguments}{The distribution arguments}

\item{value}{The query values}

\item{endpoint_suffix}{The distribution function type}
}
\value{
Vector of function values
}
\description{
Create a new distribution and query it
}
