% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{dmakedist}
\alias{dmakedist}
\title{Density of a Distribution}
\usage{
dmakedist(x, family = NULL, arguments = NULL, path = NULL, api_settings = NULL)
}
\arguments{
\item{x}{Vector of points at which to evaluate the density.}

\item{family}{A string representing the requested distribution family (optional if path is provided).}

\item{arguments}{A list containing the arguments specific to the distribution (optional if path is provided).}

\item{path}{A string representing the URL path of an existing distribution (optional if family and arguments are provided).}

\item{api_settings}{List containing API settings (optional).}
}
\value{
Vector of density values.
}
\description{
This function queries the density of a specified distribution at given points.
}
