% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_distribution}
\alias{get_distribution}
\title{Get Distribution Details}
\usage{
get_distribution(path, api_settings = NULL)
}
\arguments{
\item{path}{The URL path for the distribution to retrieve.}

\item{api_settings}{List containing API settings (optional).}
}
\value{
Parsed JSON response containing the distribution details.
}
\description{
This function retrieves the details of a distribution specified by its path.
}
