% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{initialize_api}
\alias{initialize_api}
\title{Initialize API settings}
\usage{
initialize_api(token, version = "v0")
}
\arguments{
\item{token}{Your API token as a string.}

\item{version}{API version, default is "v0".}
}
\value{
A list containing API settings.
}
\description{
This function initializes the API configuration and settings.
}
