% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{make_get_request}
\alias{make_get_request}
\title{Make an HTTP GET request}
\usage{
make_get_request(api_settings, endpoint)
}
\arguments{
\item{api_settings}{List containing base URL and token}

\item{endpoint}{API endpoint for the GET request}
}
\value{
Parsed JSON response
}
\description{
Make an HTTP GET request
}
