% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{qmakedist}
\alias{qmakedist}
\title{Quantile Function}
\usage{
qmakedist(p, family = NULL, arguments = NULL, path = NULL, api_settings = NULL)
}
\arguments{
\item{p}{Vector of probabilities at which to evaluate the quantile function.}

\item{family}{A string representing the requested distribution family (optional if path is provided).}

\item{arguments}{A list containing the arguments specific to the distribution (optional if path is provided).}

\item{path}{A string representing the URL path of an existing distribution (optional if family and arguments are provided).}

\item{api_settings}{List containing API settings.}
}
\value{
Vector of quantile values.
}
\description{
This function queries the quantile function of a specified distribution for given probabilities.
}
