% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{query_distribution}
\alias{query_distribution}
\title{Query an existing distribution}
\usage{
query_distribution(api_settings, path, value, endpoint_suffix)
}
\arguments{
\item{api_settings}{List containing API settings}

\item{path}{The URL path for an existing distribution}

\item{value}{The query values}

\item{endpoint_suffix}{The distribution function type}
}
\value{
Vector of function values
}
\description{
Query an existing distribution
}
