% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{rmakedist}
\alias{rmakedist}
\title{Random Sampling}
\usage{
rmakedist(
  size,
  family = NULL,
  arguments = NULL,
  path = NULL,
  api_settings = NULL
)
}
\arguments{
\item{size}{Integer specifying the number of samples to retrieve.}

\item{family}{A string representing the requested distribution family (optional if path is provided).}

\item{arguments}{A list containing the arguments specific to the distribution (optional if path is provided).}

\item{path}{A string representing the URL path of an existing distribution (optional if family and arguments are provided).}

\item{api_settings}{List containing API settings.}
}
\value{
Vector of random samples.
}
\description{
This function queries random samples from a specified distribution.
}
