% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate the input arguments}
\usage{
validate_arguments(path, family, arguments)
}
\arguments{
\item{path}{The URL path for an existing distribution.}

\item{family}{The distribution family.}

\item{arguments}{The distribution arguments.}
}
\value{
TRUE if validation passes, otherwise stops with an error message.
}
\description{
This function ensures that the correct set of arguments is provided.
}
