% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{dmakedist}
\alias{dmakedist}
\title{Query the density of a distribution}
\usage{
dmakedist(x, api_settings, family, arguments)
}
\arguments{
\item{x}{Vector of points at which to evaluate the density}

\item{api_settings}{List containing API settings}

\item{family}{A string representing the requested distribution family}

\item{arguments}{A list containing the arguments specific to the distribution}
}
\value{
Vector of density values
}
\description{
Query the density of a distribution
}
