% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{make_get_request}
\alias{make_get_request}
\title{Make an HTTP GET request}
\usage{
make_get_request(api_settings = NULL, endpoint)
}
\arguments{
\item{api_settings}{List containing base URL and token (optional)}

\item{endpoint}{API endpoint for the GET request}
}
\value{
Parsed JSON response
}
\description{
This function performs an HTTP GET request to a specified endpoint. If a token is provided in the API settings,
it will include an Authorization header.
}
