% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{make_post_request}
\alias{make_post_request}
\title{Make an HTTP POST request}
\usage{
make_post_request(api_settings, endpoint, body)
}
\arguments{
\item{api_settings}{List containing base URL and token}

\item{endpoint}{API endpoint for the POST request}

\item{body}{A list containing the body of the POST request}
}
\value{
Parsed JSON response
}
\description{
Make an HTTP POST request
}
