% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{pmakedist}
\alias{pmakedist}
\title{Query the cumulative distribution function of a distribution}
\usage{
pmakedist(x, api_settings, family, arguments)
}
\arguments{
\item{x}{Vector of points at which to evaluate the CDF}

\item{api_settings}{List containing API settings}

\item{family}{A string representing the requested distribution family}

\item{arguments}{A list containing the arguments specific to the distribution}
}
\value{
Vector of CDF values
}
\description{
Query the cumulative distribution function of a distribution
}
