% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{qmakedist}
\alias{qmakedist}
\title{Query the quantile function of a distribution}
\usage{
qmakedist(p, api_settings, family, arguments)
}
\arguments{
\item{p}{Vector of probabilities at which to evaluate the quantile function}

\item{api_settings}{List containing API settings}

\item{family}{A string representing the requested distribution family}

\item{arguments}{A list containing the arguments specific to the distribution}
}
\value{
Vector of quantile values
}
\description{
Query the quantile function of a distribution
}
