% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{query_distribution}
\alias{query_distribution}
\title{Create and query a distribution}
\usage{
query_distribution(value, api_settings, family, arguments, endpoint_suffix)
}
\arguments{
\item{value}{Single value specifying the size for samples or vector of x values or probabilities}

\item{api_settings}{List containing API settings}

\item{family}{A string representing the requested distribution family}

\item{arguments}{A list containing the arguments specific to the distribution}

\item{endpoint_suffix}{A string indicating the specific endpoint ('pdf', 'cdf', 'qf', or 'samples')}
}
\value{
Vector of function values or random samples
}
\description{
Create and query a distribution
}
