% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{rmakedist}
\alias{rmakedist}
\title{Query random samples from a distribution}
\usage{
rmakedist(size, api_settings, family, arguments)
}
\arguments{
\item{size}{Integer specifying the number of samples to retrieve}

\item{api_settings}{List containing API settings}

\item{family}{A string representing the requested distribution family}

\item{arguments}{A list containing the arguments specific to the distribution}
}
\value{
Vector of random samples
}
\description{
Query random samples from a distribution
}
